pyproject.toml
[project]
name = "notipusproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey Kartavtsev",email = "andrey.kartavtsev@engee.cloud"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "django-ninja (>=1.3.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "stripe (>=11.6.0,<12.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
